Add support for global deployment annotations
When secrets and configmaps change in a cluster via
https://linkerd.io/2/tasks/automatically-rotating-control-plane-tls-credentials/,
pods don't restart with the new secret/configmap information and
therefore end up using stale data.

The mechanism we use to restart deployments requires the use of
an annotation in the deployment metadata. This change enables users to
add their deployment annotations to the values.yaml.

This chart was deployed in a test cluster, and it was confirmed that
the annotations we added to our values.yaml file were propagated to the
deployments. Changing configmap/secret data resulted in deployments
restarting, as expected.

Fixes #5489

Signed-off-by: Gabrielle Madden <gabriellemadden@users.noreply.github.com>