linkerd-metrics-api bootstrap files
This is based on the branch `alpeb/grpc-viz` from #5557.

Here we complete the work started in #5554 ("Chart templates for new viz linkerd-metrics-api pod") by providing the Dockerfile, `main.go` command and the `bin/` scripts for building the new `linkerd-metrics-api` pod under the `linkerd-viz` namespace.

At the same time, we strip out of the public-api's `main.go` file the prometheus parameters and other no longer relevant bits.

Tests are not expected to pass on this one either.