experiment: replace pointer types with values to avoid nil pointer segfaults
This commit replaces pointer types with value types in Values.go to avoid
segfaults. This change should not be as easy and should need some more
changes in the templates where these values are used but hoping
it might work and avoid us adding getters which seems
a bit complex to add for every field.

If this wont work or is too complex. Lets go with Getters.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>