viz: make checks aware of prom and grafana being optional
Fixes #5618

Currently, The linkerd-viz checks fail whenever external
prometheus is being used as those checks are not aware of
prometheus and grafana being optional.

This commit fixes this by making the Prometheus and Grafana
as separate checks which are not fatal and these checks
can also be made dynamic and be ran only if those
components are available.

This commit also adds some of the missing resources checks,
especially of that of the new `metrics-api` into viz checks

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>