Remove extra three dashes from helm templates
Three dashes mean the start of the YAML document (or the end of the
directive).
If there are only comments between `---`, the document is empty.
Assume the file which include an empty document at the top of itself.
When we encrypt and decrypt it with sops, the empty document will be
converted to `{}`.
It is invalid as k8s manifest ([apiVersion not set, kind not set]).

Signed-off-by: Takumi Sue <u630868b@alumni.osaka-u.ac.jp>