Add new external-* annotations to linkerd-viz namespace
This commit adds new annotations that are added when a
external prometheus or grafana is used. Based on these
annotations, The CLI can get to know if a external instance
is being used and if the annotation is absent, that the
default instance is present.

This logic can be used by the CLi to perform checks, etc

This also cleans up some `grafanaUrl` stuff from the core
control-plane chart.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>