Remove cluster_domain. controller_log_level yaml anchors
This commit removes the usage of yaml anchors to propogate
values in the core control-plane. Instead we directly use
`default` in templates. `image_pull_policy` and `linkerd_version`
could not be removed as they are used for `debugContainer`
fields which we dont use in tempaltes but storte them direcly
in config and used by the injector and hence no way to directly
fail back in charts.

This also updates tests and golden files which expect
some of the now fields to be present

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>