Handle incompatible automount token setting In service account YAML  (#5549)
Fixes #4651

This PR adds a check in the proxy injector webhook. This checks if the concerned container has serviceaccount volume mounted. In case, the volume is not mounted, which essentially means that automountServiceAccountToken field in the service account used by the injected workload is set to false, we don't inject the sidecar proxy.

Signed-off-by: Jimil Desai <jimildesai42@gmail.com>