Stop loading cli-bin in bin/image-load
The CLI binaries are never used inside a cluster so there's no need to
load the cli-bin image.

Also, I always have `LINKERD_LOCAL_BUILD_CLI=1` in my environment to avoid
building that image to begin with (takes more than 80s in my machine
with warm docker dependencies), which caused `bin/image-load` to fail
when it attempted to load it.