prevent duplicate prometheus args
If a user specifies a `log.level` flag for linkerd-prometheus in
`prometheus.args`, the template for linkerd-prometheus will generate a
prometheus command where `--log.level` is specified twice (the first
being directly interpolated from `prometheus.logLevel` in the template),
and prometheus will crash-loop because its flags parser does not allow
duplicate flags that are not arrays.

Only fill in the `--log.level` flag from `.Values.prometheus.logLevel`
if the `log.level` key is _not_ present in `.Values.prometheus.args`

Added a test case.

Signed-off-by: Nathan J. Mehl <n@oden.io>