Warn users about missing tap configuration
This change modifies the dashboard so that it stops displaying an error
message when viewing tap streams from resources that don't have tap
enabled and instead, shows a warning and also guides the user on what to
do to fix it.

This change also includes updating the `<ExpansionPanel>` component
since we were getting deprecation warnings from Material-UI. The
component has been renamed to an accordion.

Fixes #5831

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>