Reenable disabled integration tests
Reenable the `upgrade-test` and tracing integration tests. The former
was disabled because in edge-21.2.4 the viz install wasn't waiting for
the core components to be ready, and the latter was disabled because for
some unknown reason the `jaeger-webhook` image couldn't be loaded.
In our latest tries we haven't seen that problem, so here it goes again
:cross-fingers: